+++ src/frontend/src/components/Smallprofile/index.css.ts
